+++ src/core/connected_client.rs
@@ -1,0 +1,107 @@
+use bevy::{
+    ecs::{component::ComponentId, world::DeferredWorld},
+    prelude::*,
+    utils::HashMap,
+};
+
+/// Marker for a connected client.
+///
+/// Backends should spawn and despawn entities with this component on connect and disconnect.
+///
+/// `Entity` is used an identifier to refer to a client.
+///
+/// <div class="warning">
+///
+/// Entities with this components should be spawned and despawned only from the messaging backend.
+///
+/// </div>
+///
+/// See also [`ReplicatedClient`](crate::server::ReplicatedClient).
+#[derive(Component, Reflect)]
+#[component(on_add = on_client_add, on_remove = on_client_remove)]
+#[require(Name(|| Name::new("Connected client")), NetworkStats)]
+pub struct ConnectedClient {
+    id: ClientId,
+}
+
+impl ConnectedClient {
+    /// Creates a new instance with a backend-provided ID.
+    pub fn new(id: u64) -> Self {
+        Self {
+            id: ClientId::new(id),
+        }
+    }
+
+    /// Returns client ID provided by backend.
+    ///
+    /// Unlike [`Entity`], it's a persistent identifier that could
+    /// be used to identify the same client after reconnects.
+    ///
+    /// See also [`ClientIdMap`].
+    pub fn id(&self) -> ClientId {
+        self.id
+    }
+}
+
+fn on_client_add(mut world: DeferredWorld, entity: Entity, _id: ComponentId) {
+    let connected_client = world.get::<ConnectedClient>(entity).unwrap();
+    let client_id = connected_client.id;
+    let mut map = world.resource_mut::<ClientIdMap>();
+    if let Some(old_entity) = map.0.insert(client_id, entity) {
+        error!("backend-provided `{client_id:?}` that was already mapped to client `{old_entity}`");
+    }
+}
+
+fn on_client_remove(mut world: DeferredWorld, entity: Entity, _id: ComponentId) {
+    let connected_client = world.get::<ConnectedClient>(entity).unwrap();
+    let client_id = connected_client.id;
+    let mut map = world.resource_mut::<ClientIdMap>();
+    map.0.remove(&client_id);
+}
+
+/// Maps [`ConnectedClient::id`] to associated entity.
+///
+/// Automatically updated on clients spawns and despawns.
+#[derive(Resource, Default, Deref)]
+pub struct ClientIdMap(HashMap<ClientId, Entity>);
+
+/// Unique client ID provided by a messaging backend.
+#[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, Ord, PartialOrd, Reflect)]
+pub struct ClientId(u64);
+
+impl ClientId {
+    /// Creates a new ID wrapping the given value.
+    pub const fn new(value: u64) -> Self {
+        Self(value)
+    }
+
+    /// Gets the value of this ID.
+    pub fn get(&self) -> u64 {
+        self.0
+    }
+}
+
+/// Statistic associated with [`RepliconClient`](super::replicon_client::RepliconClient) or
+/// [`ConnectedClient`].
+///
+/// All values can be zero if not provided by the backend.
+///
+/// <div class="warning">
+///
+/// Should only be modified from the messaging backend.
+///
+/// </div>
+#[derive(Component, Debug, Clone, Copy, Default, Reflect)]
+pub struct NetworkStats {
+    /// Round-time trip in seconds for the connection.
+    pub rtt: f64,
+
+    /// Packet loss % for the connection.
+    pub packet_loss: f64,
+
+    /// Bytes sent per second for the connection.
+    pub sent_bps: f64,
+
+    /// Bytes received per second for the connection.
+    pub received_bps: f64,
+}